requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.3.1
albumentations==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==21.4b2
blinker==1.4
blis==0.7.11
boto3==1.29.4
botocore==1.32.4
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
ci-info==0.3.0
cleo==2.1.0
click==8.0.3
cloud-init==23.3.1
cloudpickle==3.0.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
configobj==5.0.6
configparser==6.0.0
constantly==15.1.0
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
datasets==2.11.0
dbus-python==1.2.18
Deprecated==1.2.14
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
dill==0.3.4
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docstring-parser==0.15
dulwich==0.21.6
effdet==0.4.1
einops==0.7.0
emoji==2.2.0
entmax==1.1
etelemetry==0.3.1
evdev==1.6.1
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
fitz==0.0.1.dev2
fonttools==4.45.0
frozenlist==1.4.0
fsspec==2023.3.0
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-api-core==1.34.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.61.0
GPUtil==1.4.0
grpcio==1.59.3
grpcio-status==1.48.2
gyp==0.1
h11==0.14.0
htmlwebshot==0.1.0
httplib2==0.20.2
httptools==0.6.1
huggingface-hub==0.17.3
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.3
imageio==2.33.0
img2pdf==0.5.0
imgkit==1.2.3
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==21.3.0
installer==0.7.0
iopath==0.1.9
isodate==0.6.1
jaraco.classes==3.3.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.3.2
jsonargparse==4.27.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=4d002dcde70b52d519cafff4dc0008696c40cff1c9184a531b40c7b45905be6b
keyring==24.3.0
kiwisolver==1.4.5
langcodes==3.3.0
latext==0.0.7
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
Levenshtein==0.23.0
lightning==2.1.2
lightning-utilities==0.10.0
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
lz4==4.3.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.12
munch==4.0.0
murmurhash==1.0.10
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
nougat-ocr==0.1.17
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.0
pathlib==1.0.1
pathspec==0.11.2
pathy==0.10.3
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.8.0
pika==1.3.2
pikepdf==8.7.1
Pillow==9.5.0
pix2tex==0.1.2
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.18.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
portalocker==2.8.2
preshed==3.0.9
protobuf==3.20.3
prov==2.0.0
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==12.0.0
pyarrow-hotfix==0.5
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.6.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pynput==1.7.6
pyOpenSSL==23.3.0
pyparsing==2.4.7
pyparted==3.11.7
pypdf==3.17.1
PyPDF2==3.0.1
pypdfium2==4.24.0
pyproject_hooks==1.0.0
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
PyQt6-WebEngine-Qt6==6.6.0
pyrsistent==0.18.1
pyserial==3.5
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.0
python-dotenv==1.0.0
python-Levenshtein==0.23.0
python-magic==0.4.24
python-xlib==0.33
pytorch-lightning==2.1.2
pytz==2022.1
pyxnat==1.6
PyYAML==5.4.1
qudida==0.0.4
rapidfuzz==3.5.2
rdflib==7.0.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
sconf==0.2.5
screen-resolution-extra==0.0.0
screeninfo==0.8.1
SecretStorage==3.3.1
selenium==4.15.2
sentencepiece==0.1.99
service-identity==18.1.0
shellingham==1.5.4
shiboken6==6.6.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snakeviz==2.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.5
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.11
streamlit==1.28.2
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
tesseract==0.1.3
texttable==1.6.4
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.5.4
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1+cu118
torchmetrics==1.2.0
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traits==6.3.2
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
triton==2.1.0
trove-classifiers==2023.11.22
Twisted==22.1.0
typer==0.7.0
typeshed-client==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unzip==1.0.0
urllib3==1.26.5
validators==0.22.0
virtualenv==20.24.7
wadllib==1.3.6
WALinuxAgent==2.2.46
wasabi==1.1.2
watchdog==3.0.0
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wkhtmltopdf==0.2
wrapt==1.16.0
wsproto==1.2.0
x-transformers==0.15.0
xkit==0.0.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.3
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.22.0
